valor era u32 agr eh u64, pq no prog kernel era u64

+++ bib/teste_bib.h
@@ -22,6 +22,6 @@ int remove_ebpf(char caminho_prog[],struct info_ebpf *infos);
 
 
 int le_mapa(struct info_ebpf *infos);
-int atualiza_mapa(char caminho_prog[], char nome_mapa[], __u32 *valor, struct info_ebpf *infos);
+int atualiza_mapa(char caminho_prog[], char nome_mapa[], __u64 *valor, struct info_ebpf *infos);
 
 #endif // MYLIB_H
